requirements.txt

streamlit
gspread
google-auth
google-generativeai
pandas
